LPD-58926 Move Services folder to common
