+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/common/services/PicklistService.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {Options, Picklist} from '../types/Picklist';
+import {Options, Picklist} from '../../types/Picklist';
 import ApiHelper from './ApiHelper';
 
 async function createPicklist({
