+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/common/services/VocabularyService.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {IVocabulary} from '../types/IVocabulary';
+import {IVocabulary} from '../../types/IVocabulary';
 import ApiHelper from './ApiHelper';
 
 async function createVocabulary(vocabulary: IVocabulary) {
