+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/common/types/Space.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {LogoColor} from '../main/components/SpaceSticker';
+import {LogoColor} from '../../main/components/SpaceSticker';
 
 export type Space = {
 	creatorUserId: string;
