+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/FDSPropsTransformer/cell_renderers/StructureScopeRenderer.tsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-import SpaceService from '../../../services/SpaceService';
+import SpaceService from '../../../common/services/SpaceService';
 import SpaceSticker, {LogoColor} from '../../components/SpaceSticker';
 
 const {useEffect, useState} = React;
