+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/categories/components/MoveCategoryModalContent.tsx
@@ -5,7 +5,7 @@
 
 import React, {useEffect, useState} from 'react';
 
-import ApiHelper from '../../../../services/ApiHelper';
+import ApiHelper from '../../../../common/services/ApiHelper';
 import MoveCategoryTreeView from './MoveCategoryTreeView';
 
 export const FETCH_URLS = {
