+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/categories/components/MoveCategoryTreeView.tsx
@@ -12,7 +12,7 @@ import {openToast} from 'frontend-js-components-web';
 import {sub} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
-import ApiHelper from '../../../../services/ApiHelper';
+import ApiHelper from '../../../../common/services/ApiHelper';
 import {FETCH_URLS} from './MoveCategoryModalContent';
 
 /**
