+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/vocabularies/EditVocabulary.tsx
@@ -11,8 +11,8 @@ import {navigate, sub} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
 import Toolbar from '../../../common/components/Toolbar';
-import CategorizationPermissionService from '../../../services/CategorizationPermissionService';
-import VocabularyService from '../../../services/VocabularyService';
+import CategorizationPermissionService from '../../../common/services/CategorizationPermissionService';
+import VocabularyService from '../../../common/services/VocabularyService';
 import {IVocabulary} from '../../../types/IVocabulary';
 import {IPermissionItem} from '../../components/forms/PermissionsTable';
 import {
