+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/MultipleFileUploader.tsx
@@ -11,12 +11,12 @@ import {sub} from 'frontend-js-web';
 import React, {useState} from 'react';
 import {useDropzone} from 'react-dropzone';
 
-import ApiHelper from '../../services/ApiHelper';
 import DragZoneBackground from './DragZoneBackground';
 import {LoadingMessage} from './LoadingMessage';
 import {FieldPicker} from './forms';
 
 import '../../../css/components/MultipleFileUploader.scss';
+import ApiHelper from '../../common/services/ApiHelper';
 import {AssetLibrary} from '../../types/AssetLibrary';
 
 interface FileData {
