+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/dashboard/components/AllCategoriesDropdown.tsx
@@ -5,7 +5,7 @@
 
 import React, {useContext, useState} from 'react';
 
-import ApiHelper from '../../../services/ApiHelper';
+import ApiHelper from '../../../common/services/ApiHelper';
 import {ViewDashboardContext} from '../ViewDashboardContext';
 import {buildQueryString} from '../utils/buildQueryString';
 import {FilterDropdown} from './FilterDropdown';
