+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/spaces/SpaceGeneralSettings.tsx
@@ -11,8 +11,8 @@ import {openToast, useId} from 'frontend-js-components-web';
 import {navigate} from 'frontend-js-web';
 import React from 'react';
 
+import SpaceService from '../../common/services/SpaceService';
 import focusInvalidElement from '../../common/utils/focusInvalidElement';
-import SpaceService from '../../services/SpaceService';
 import {Space} from '../../types/Space';
 import {LogoColor} from '../components/SpaceSticker';
 import {FieldText} from '../components/forms';
