+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/contexts/CacheContext.tsx
@@ -14,8 +14,8 @@ import React, {
 	useState,
 } from 'react';
 
-import PicklistService from '../../services/PicklistService';
-import SpaceService from '../../services/SpaceService';
+import PicklistService from '../../common/services/PicklistService';
+import SpaceService from '../../common/services/SpaceService';
 import {Picklist} from '../../types/Picklist';
 import {Space} from '../../types/Space';
 import StructureService from '../services/StructureService';
