+++ modules/apps/site/site-cms-site-initializer/test/js/main/spaces/AddSpaceMembers.test.tsx
@@ -8,11 +8,11 @@ import {act, render, screen, waitFor, within} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
+import SpaceService from '../../../../src/main/resources/META-INF/resources/js/common/services/SpaceService';
 import {
 	AddSpaceMembers,
 	AddSpaceMembersProps,
 } from '../../../../src/main/resources/META-INF/resources/js/main/spaces/AddSpaceMembers';
-import SpaceService from '../../../../src/main/resources/META-INF/resources/js/services/SpaceService';
 import {Space} from '../../../../src/main/resources/META-INF/resources/js/types/Space';
 import {
 	UserAccount,
