+++ modules/apps/site/site-cms-site-initializer/test/js/structure_builder/contexts/CacheContext.test.tsx
@@ -9,14 +9,14 @@ import '@testing-library/jest-dom/extend-expect';
 import {render, waitFor} from '@testing-library/react';
 import React from 'react';
 
-import SpaceService from '../../../../src/main/resources/META-INF/resources/js/services/SpaceService';
+import SpaceService from '../../../../src/main/resources/META-INF/resources/js/common/services/SpaceService';
 import {useCache} from '../../../../src/main/resources/META-INF/resources/js/structure_builder/contexts/CacheContext';
 import {Space} from '../../../../src/main/resources/META-INF/resources/js/types/Space';
 import {MockCacheProvider} from '../mocks/MockCacheProvider';
 import {MockStateProvider} from '../mocks/MockStateProvider';
 
 jest.mock(
-	'../../../../src/main/resources/META-INF/resources/js/services/SpaceService',
+	'../../../../src/main/resources/META-INF/resources/js/common/services/SpaceService',
 	() => ({
 		getSpaces: jest.fn(() => Promise.resolve([])),
 	})
