+++ modules/apps/site/site-cms-site-initializer/test/js/structure_builder/mocks/MockCacheProvider.tsx
@@ -5,8 +5,8 @@
 
 import React, {ReactNode} from 'react';
 
-import PicklistService from '../../../../src/main/resources/META-INF/resources/js/services/PicklistService';
-import SpaceService from '../../../../src/main/resources/META-INF/resources/js/services/SpaceService';
+import PicklistService from '../../../../src/main/resources/META-INF/resources/js/common/services/PicklistService';
+import SpaceService from '../../../../src/main/resources/META-INF/resources/js/common/services/SpaceService';
 import {CacheContext} from '../../../../src/main/resources/META-INF/resources/js/structure_builder/contexts/CacheContext';
 import StructureService from '../../../../src/main/resources/META-INF/resources/js/structure_builder/services/StructureService';
 import {Structures} from '../../../../src/main/resources/META-INF/resources/js/structure_builder/types/Structure';
